+++ CVSystem/Proc/Event.cpp
@@ -41,7 +41,7 @@ namespace dxlib {
         }
 
         EventItem* ptr = (EventItem*)memEventData;
-        for (size_t i = 0; i < 64; i++) { //最多支持64条事件吧
+        for (int i = 0; i < 64; i++) { //最多支持64条事件吧
             if (ptr[i].flag != 0) {
                 ptr[i].flag = 0;//标记回0
